# Conflicts:
#	app/code/Magento/CatalogImportExport/Model/Import/Product.php
#	app/code/Magento/CatalogSearch/Model/Adapter/Mysql/Aggregation/DataProvider.php
#	app/code/Magento/CatalogSearch/Test/Unit/Model/Adapter/Mysql/Aggregation/DataProviderTest.php
#	app/code/Magento/Store/Setup/InstallSchema.php
#	app/code/Magento/Ui/view/base/web/js/dynamic-rows/dynamic-rows.js
#	app/etc/di.xml
#	composer.lock

+++ app/code/Magento/UrlRewrite/Model/Storage/DbStorage.php
@@ -147,7 +147,10 @@ class DbStorage extends AbstractStorage
     {
         foreach ($this->createFilterDataBasedOnUrls($urls) as $type => $urlData) {
             $urlData[UrlRewrite::ENTITY_TYPE] = $type;
-            $this->deleteByData($urlData);
+            // prevent query locking in a case when nothing to delete
+            if ($this->connection->fetchRow($this->prepareSelect($urlData))) {
+                $this->deleteByData($urlData);
+            }
         }
         $data = [];
         foreach ($urls as $url) {
